pyproject.toml
[project]
name = "obsidian-rag"
version = "1.0.0"
description = "RAG with Gemini and ChromaDB for Obsidian notes with a TUI interface"
requires-python = ">=3.13"
dependencies = [
    "langchain[google-genai]>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.1.0",
    "langchain-google-genai>=4.1.2",
    "unstructured>=0.18.21",
    "markdown>=3.10",
    "langchain-chroma>=1.1.0",
    "rich>=14.2.0",
    "prompt-toolkit>=3.0.52",
    "pylatexenc>=2.10",
    "python-dotenv>=1.1.0",
    "redis>=5.0.0",
    "psycopg[binary]>=3.2.1",
]

[project.scripts]
chat = "cli.chat:main"
build_index = "cli.build_index:main"

[tool.uv]
package = true
